Fix sdbdpsdump format of INVALIDATECATA
git-svn-id: http://192.168.20.11/sequoiadb/trunk@29478 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/dps/dpsLogRecord.cpp
@@ -924,11 +924,11 @@ namespace engine
                                  " Type    : %s(%d)"OSS_NEWLINE,
                                  "INVALIDATE CATA", LOG_TYPE_INVALIDATE_CATA ) ;
             dpsLogRecord::iterator itrType, itrCL, itrIX ;
-            UINT32 invType = 0 ;
+            UINT8 invType = 0 ;
             itrType = this->find( DPS_LOG_INVALIDCATA_TYPE ) ;
             if ( itrType.valid() )
             {
-               invType = *( ( UINT32 * )( itrType.value() ) ) ;
+               invType = *( ( UINT8 * )( itrType.value() ) ) ;
                len += ossSnprintf ( outBuf + len, outSize - len,
                                     " InvType : 0x%02x (%u)"OSS_NEWLINE,
                                     invType, invType ) ;
